adds missing steps
Hello,

I wasn't able to get the app to compile by using the exact steps in the readme. This PR changes it so that it at compiles. It seems like the Gleam Std Lib needs to be compiled manually with rebar first and then any user gleam code gets compiled with the `iex -S mix`.

Also thanks so much for putting this together!